Initial structure (#1)
Set up some initial structure to the project that includes the
following:
- empty public header per platform
- empty static library per platform
- prepare-machine.ps1 tool
- build.ps1 tool
- CI pipeline that builds the library for all platforms
- dependabot

All of this was taken from https://github.com/microsoft/msquic and
trimmed down. There may still be some unnecessary stuff that is
lingering (especially the cmake stuff which is foreign to me), but I
think this is a reasonable start.